\name{E1}
\alias{E1}
\title{An edge matrix representing a similarity graph}
\description{This is the variable name for an edge matrix in the "Example" data. It is constructed from a sequence of n=200 observations with a change in mean at t = 100. E1 is a matrix with the number of rows the number of edges in the similarity graph and 2 columns.  Each row contains the node indices of an edge.
}

