\name{E3}
\alias{E3}
\title{An edge matrix representing a similarity graph}
\description{This is the variable name for an edge matrix in the "Example" data. It is constructed from a sequence of n=200 observations with a change in mean and variance starting at t = 145. E3 is a matrix with the number of rows the number of edges in the similarity graph and 2 columns.  Each row contains the node indices of an edge.
}

