\name{E5}
\alias{E5}
\title{An edge matrix representing a similiarity graph}
\description{This is the variable name for an edge matrix in the "Example" data. It is constructed from a sequence of n=200 observations with a change in mean on the interval t= 155 to t=185. E5 is a matrix with the number of rows the number of edges in the similarity graph and 2 columns.  Each row contains the node indices of an edge.
}
